pyproject.toml
[project]
name = "qa-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-anthropic>=1.2.0",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.3",
    "playwright>=1.56.0",
    "python-dotenv>=1.2.1",
]
